+++ CXPMobileSDK/mfe3e6a-internal-navigation/DemoWidgetsAssets/assets/backbase/static/features/[BBHOST]/module-places/scripts/main.js
@@ -6,9 +6,8 @@ define(function (require, exports, module) {
     var core = require('core');
     var base = require('base');
     var ui = require('ui');
-
-
     require('./libs/angular-gm');
+
     // requirejs async plugin
     // Is ignored by webpack
     require('async!http://maps.google.com/maps/api/js?libraries=places&sensor=true&');
@@ -20,6 +19,7 @@ define(function (require, exports, module) {
     ];
 
     module.exports = base.createModule(module.name, deps)
+        .config( require('./config') )
         .constant( require('./utils') )
         .service( require('./services') )
         .directive( require('./places-autocomplete') );