+++ CXPMobileSDK/mfe3e6a-internal-navigation/DemoWidgetsAssets/assets/backbase/static/widgets/[BBHOST]/widget-notifications/scripts/constants.js
@@ -3,6 +3,7 @@ define(function(require, exports, module) {
 
     exports.widgetPrefs = {
         NOTIFICATIONS_ENDPOINT: 'notificationsEndpoint',
+        IGNORE_ENROLLMENT_ENDPOINT: 'ignoreEnrollmentEndpoint',
         CLOSE_NOTIFICATION_ENDPOINT: 'closeNotificationEndpoint',
         POLL_INTERVAL: 'pollInterval',
         ALLOW_PUBSUB: 'allowPubsub'